backend/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
sentence-transformers==2.2.2
faiss-cpu==1.13.2
python-dotenv==1.0.0
pandas==2.2.2
openpyxl==3.1.2
requests
